Use int64 for eamil.Throttle related values
After working on commit 0e05d3cbf51689ed966aba8dee64c68ed87d5adb,
whereby all the DynamoDB item counts are int64, it seemed to make sense
to use int64 for the Throttle.

+++ email/mailer.go
@@ -11,7 +11,7 @@ import (
 )
 
 type Mailer interface {
-	BulkCapacityAvailable(ctx context.Context, numToSend int) error
+	BulkCapacityAvailable(ctx context.Context, numToSend int64) error
 
 	Send(
 		ctx context.Context, recipient string, msg []byte,
@@ -25,7 +25,7 @@ type SesMailer struct {
 }
 
 func (mailer *SesMailer) BulkCapacityAvailable(
-	ctx context.Context, numToSend int,
+	ctx context.Context, numToSend int64,
 ) error {
 	return mailer.Throttle.BulkCapacityAvailable(ctx, numToSend)
 }
